third_party/blink/web_tests/wpt_internal/sanitizer-api/sanitizer-sanitize.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/testcases.sub.js"></script>
</head>

<body>
<script>
    function getString(fragment) {
      d = document.createElement("div");
      d.appendChild(fragment);
      return d.innerHTML;
    }

    test(t => {
      let s = new Sanitizer({});
      assert_throws_js(TypeError, _ => s.sanitize());
    }, "SanitizerAPI sanitize function without argument should throw an error.");

    test(t => {
      let s = new Sanitizer({});
      fragment = s.sanitize(null);
      assert_true(fragment instanceof DocumentFragment);
      assert_equals(getString(fragment), "null");
    }, "SanitizerAPI sanitize function for null.");

    testcases.forEach(c => test(t => {
        let s = new Sanitizer(c.config_input);
        fragment = s.sanitize(c.value);
        assert_true(fragment instanceof DocumentFragment);
        assert_equals(getString(fragment), c.result);
    }, "SanitizerAPI with config: " + c.message + ", sanitize from string function for " + c.message));

    testcases.forEach(c => test(t => {
        let s = new Sanitizer(c.config_input);
        var dom = document.implementation.createHTMLDocument();
        dom.documentElement.innerHTML = c.value;
        fragment = s.sanitize(dom);
        assert_true(fragment instanceof DocumentFragment);

        let result = getString(fragment);
        // Remove <html> and </html>
        if (c.message != "document" && result.slice(0,12) == "<html><body>" && result.slice(result.length-14, result.length) == "</body></html>")
          result = result.substr(12, result.length - 26);
        if (c.message == "document") {
          assert_equals(result, "<html><body>test</body></html>");
        } else {
          assert_equals(result, c.result);
        }
    }, "SanitizerAPI with config: " + c.message + ", sanitize from document function for " + c.message));

    testcases.forEach(c => test(t => {
        let s = new Sanitizer(c.config_input);
        let tpl = document.createElement("template");
        tpl.innerHTML = c.value;
        fragment = s.sanitize(tpl.content);
        assert_true(fragment instanceof DocumentFragment);
        assert_equals(getString(fragment), c.result);
    }, "SanitizerAPI with config: " + c.message + ", sanitize from document fragment function for " + c.message));

    test(t => {
      let s = new Sanitizer();
      let policy = trustedTypes.createPolicy("myPolicy", {createHTML: s=>s+1});
      let html_string = policy.createHTML("testHTML");
      fragment = s.sanitize(html_string);
      assert_true(fragment instanceof DocumentFragment);
      assert_equals(getString(fragment), "testHTML1");
    }, "SanitizerAPI sanitize from TrustedHTML.");

    test(t => {
      let s = new Sanitizer();
      let policy = trustedTypes.createPolicy("default", {createHTML: s=>s+2});
      let html_string = policy.createHTML("testHTML");
      fragment = s.sanitize(html_string);
      assert_true(fragment instanceof DocumentFragment);
      assert_equals(getString(fragment), "testHTML2");
    }, "SanitizerAPI sanitize from string with default policy.");
</script>
</body>
</html>
